requirements.txt
langchain-core
langchain-community
langchain-groq
langgraph
fastapi
numpy
pandas
python-dotenv
huggingface-hub
transformers
sentence-transformers
faiss-cpu
bs4
langchain-huggingface
langchain-classic
nltk
tensorflow
tf-keras
torch
torchvision
torchaudio
-f https://download.pytorch.org/whl/cpu/torch_stable.html
datasets
peft
accelerate
bitsandbytes
beautifulsoup4
uvicorn
jinja2
tavily-python